+++ kernels/TileMethod.cl
@@ -1,6 +1,6 @@
 #define TS 32
 
-__kernel void vector_add(__global float* A, __global float* B, __global float* C, 
+__kernel void vector_add(__global double* A, __global double* B, __global double* C, 
 						const int M, const int N, const int K) 
 {
     const int row = get_local_id(0);
@@ -8,10 +8,10 @@ __kernel void vector_add(__global float* A, __global float* B, __global float* C
     const int global_row = TS * get_group_id(0) + row;
     const int global_col = TS * get_group_id(1) + col;
 
-    __local float Aloc[TS][TS];
-    __local float Bloc[TS][TS];
+    __local double Aloc[TS][TS];
+    __local double Bloc[TS][TS];
 
-    float sum = 0.0f;
+    double sum = 0.0f;
 
     const int num_tiles = K / TS;
     for (int t = 0; t < num_tiles; t++)
